web: rename RootAuthMethods to RootUser
